refactor: share default link/anchor resolution across HTML contexts
Cubic finding (PR 149, src/formats/mhtml.rs): MhtmlCtx::link_target and
MhtmlCtx::anchor_id duplicated StandaloneCtx verbatim. Move both
implementations into HtmlCtx as default trait methods so only the
image_source hook (which genuinely differs per frontend) and
frontends with scoped link semantics (EPUB chapters) carry their own
implementations. Behavior is unchanged for every frontend.
